Fix TranslateGemma crash handling and improve user experience
Three improvements to address user-reported issues:

1. **Better error messages for "failing to create engine"**
   - Added model file full paths and sizes in GB for clarity
   - Separated exception handling for UnsatisfiedLinkError (GPU libraries missing)
   - Added OutOfMemoryError handling with device memory context
   - Improved conversation/flow creation error messages with actionable guidance
   - All error messages now provide specific context for debugging

2. **Increase error toast visibility duration**
   - Changed SnackbarDuration from Short to Long in TranslationPage
   - Error messages now remain visible for 10 seconds instead of 4

3. **Set Hebrew to English as default language pair**
   - Changed default source language from Auto to Hebrew
   - TranslationModel now defaults to Hebrew→English translation

+++ app/src/main/java/com/barakplasma/privateaitranslate/ui/screens/TranslationPage.kt
@@ -145,7 +145,7 @@ fun TranslationPage(
                     val result = snackbarHostState.showSnackbar(
                         message = errorMessage,
                         actionLabel = context.getString(R.string.options),
-                        duration = SnackbarDuration.Short
+                        duration = SnackbarDuration.Long
                     )
                     when (result) {
                         SnackbarResult.ActionPerformed -> navHostController.navigate(Destination.Settings.route)
